fix(codex): keep client defaults browser-safe
